feat(theme): unify dark and light mode handling
